DON'T BE AFRAID :D. EXPEDITION SERVICE DOESN'T WORK YET. Refactor expedition and seat services: update createExpedition method to include companyId, add seat generation logic, and implement repository queries for city and seat management.

+++ backend/expedition-service/src/main/java/com/shubilet/expedition_service/services/SeatService.java
@@ -7,7 +7,7 @@ import com.shubilet.expedition_service.dataTransferObjects.responses.SeatForCust
 
 public interface SeatService {
     
-    public void generateSeats(long expeditionId, int capacity);
+    public void generateSeats(int expeditionId, int capacity);
 
     public List<SeatForCustomerDTO> getAvailableSeats(int expeditionId);
